Move emoji utils out of the auto-generated file
