+++ components/emoji_picker/emoji_picker.jsx
@@ -7,6 +7,7 @@ import {FormattedMessage} from 'react-intl';
 import throttle from 'lodash/throttle';
 
 import * as Emoji from 'utils/emoji.jsx';
+import {sortEmojis} from 'utils/emoji_utils.jsx';
 import {t} from 'utils/i18n';
 import imgTrans from 'images/img_trans.gif';
 
@@ -383,13 +384,13 @@ export default class EmojiPicker extends React.PureComponent {
             emojiArray.push(emoji);
         });
 
-        const sortEmojis = (a, b) => {
-            return Emoji.sortEmojis(a, b, this.state.filter);
+        const sortEmojisHelper = (a, b) => {
+            return sortEmojis(a, b, this.state.filter);
         };
 
-        recentEmojis.sort(sortEmojis);
+        recentEmojis.sort(sortEmojisHelper);
 
-        emojisMinusRecent.sort(sortEmojis);
+        emojisMinusRecent.sort(sortEmojisHelper);
 
         return [
             ...recentEmojis,
